uncrouch data and pivot data is better organized
Now all uncrouch info is in object "uncrouch", and pivot info is in object "pivot".
Likewise, dash zone info is in object "dashzone" and crouch zone info is in "crouchZone"
Data will be easier to pass, and both objects share property names which means
that I can use the same code to operate them. (and i will)