Refactor the deadzoneExits
now instead of reflecting an action (deadzoneExit.[...].did) we describe a state (outOfDeadzone.[...].is).

It's better this way because ideally we should treat it the same as dash zones, sdi zones and crouch zone, which don't create timing lockouts after completion. not by themselves.

To illustrate this: imagine there was no dashzone history array but we keep dashZone.saved (this scenario meets dashZone.historyLength == 1, dashZone.hist[1] == dashZone.saved). we can delete the information of a previous dash if we bring the stick to neutral, as now this information of stick neutral will take the place of the dash.

We can't do the same thing with pivot history. if we performed a recent empty pivot, it's saved. If afterwards, we move the stick, say, to perform uptilt, we detect that this wasn't the completion of a pivot, but we won't go and pop the object referenced by pivot.saved.

we refrain to writing the failed pivot check in history in order to be able to keep reading the last pivot timestamp and uphold the timing lockout. of note is that recent pivot completions can overwrite active old ones.

This ideal case doesn't accurately represent the current logic written on the script, but I will fix it (the scripts) in the following commits.