implemented a new outOfDeadzone object queue
the simplest method so far of making a queue that stores the timestamps of the earliest encounter of any particular value, is this:

make array addedToQueue (to track if we already had our first encounter with a particular value) and queueTimestamp (stores the timestamps of said first encounters)

those two start empty and we add timestamps to values we encounter within a key multipress timeframe.

if we encounter a value again in a short period of time, we retrieve the timestamp from queueTimestamp[value].

when the multipress timelimit expires we flush the queue and save the last unsaved value + accompanying timestamp.

i tested this concept with outOfDeadzone.up and outOfDeadzone.down. it works - shortly i will update the queue manipulation code of the other history keeping objects