Changed the queue system for crouchZone saving
I implemented storing crouchZoneHistoryEntry objects in the queue of the crouchZone. I changed the way outOfDeadzone.[direction] queue works compared to the previous commit. In both crouchZone and outOfDeadzone[direction] I save object addresses in the queue and I retrieve them if the stick re-enters a zone again and the script hasn't flushed the queue