fixed a bug where assigning duplicate hotkeys via back and forth worked
to do it you needed to attempt this with two different controls that are not the one you want to copy from. Assigning a duplicate to the first gets denied, quickly assigning the second one got denied too, but then quickly assigning the first was allowed.
Thread interruption was the culprit, so I created a free to use thread via setTimer.