Launch mode bug fix and massive refactoring.
- fixed input toggle recall on launch not working because of a typo. ControlsEnabledRecall -> EnabledControlsRecall is the proper ini key.

Refactoring

If the coder wants to write a function call, now he/she will have a little more work to do.
The coder that wants to call a function will need to enter PROCESSED data as parameters instead of just passing aX and aY and have the function extract the information of interest from the coordinates. Even if it makes a function call harder to write, this is easier on testing and maybe it's also slightly faster. (Apparently for AutoHotkey, the more lines of code, the slower the script runs.) This approach to refactoring is what I used in last commits, and finally I took time to explain it. These are some ways that this approach paid off for this particular commit:

- merge the getCurrentOutOfDeadzoneInfo_up and getCurrentOutOfDeadzoneInfo_down into a single function
- directionDeadzoneHistoryObject doesn't depend on a different Func object for each of the 4 directions anymore. as such, class extensions upDeadzoneHistoryObject and downDeadzoneHistoryObject do not need to be a thing anymore
- getCurrentDashZoneInfo() doesn't need an aX parameter passed to it
- getDidPivot() doesn't need an aX parameter passed to it
- getUncrouchDid() doesn't need an aY parameter passed to it

Other changes:
- vscode intellisense is broken for function parameters in ahk v1 so I need to do things like this:
pivot.storeInfoBeforeMultipressEnds_pivot() as opposed to pivot.storeInfoBeforeMultipressEnds_pivot(),
pivot.pivotNerfSearch() as opposed to pivot.nerfSearch(),

and, reordering the function parameters, so that the ones that we usually need to compute (process) with ANOTHER function are placed last. Such parameters are usually the output of functions that work with aX and aY.

So, in most cases*, things go like this:
functionCall(history object that holds previous processed data ... , ... processed data of current coordinates aX, aY)

* outOfDeadzone, dashZone, crouchZone, pivot, uncrouch, limitOutputs() maybe