Pivot detection consistency upon regressing to queue timestamp
i did away with the dashZone history entry stale flags and checkHistoryEntryStaleness(). Now we observe if an input sequence timing is stale, in pivotTimingCheck(), by comparing it to the current dashzone timestamp (that is either saved.timestamp, queue[thing].timestamp or currentTimeMS) instead of only comparing them to currentTimeMS as we did in the old checkHistoryEntryStaleness(). Here's a malfunction of the old design: Suppose we stop our dash at the very last moment that allows us to perform a pivot (going solely off of TIMESTALE_PIVOT_INPUTSEQUENCE). Then in a single multiple keypress/lift action we get away from dash stop range and return to it (all in less than 3ms since the initial dash stop). Now out input wont be detected as a pivot. It will be classified as a stale sequence instead.
With our change we avoid that inconsistency. If the pivot input was detected once before with a timestamp that is in the queue, it will be detected again if we retrieve that timestamp.

pivotNerfSearch() and uncrouchNerfSearch() use a greatly simplified logic.

Other refactoring.