Changing where pivot and uncrouch info gets allocated
Every dash history entry will have its associated pivot property where we indicate if the player finished a pivot with that dash zone change.
I implemented the same system for dashZone and crouchZone, even if it was not as necessary for the latter, but for consistency I did it. Every crouch zone history entry now has a uncrouch property.